+++ src/basil/error.clj
@@ -7,31 +7,31 @@
 ;; === Rendering error messages ===
 
 
-(defn text-error
-  "Render error message as regular text"
-  [text] {:post [(types/error-text? %)]
-          :pre  [(types/error-text? text)]}
-  (assoc text :text (str "ERROR: " (:text text))))
-
-
-(defn html-error
-  "Render error message as HTML snippet"
-  [text] {:post [(types/error-text? %)]
-          :pre  [(types/error-text? text)]}
-  (assoc
-    text :text
-    (format "<pre style='color: white; background-color: red'>ERROR: %s</pre>"
-            (util/html-entities (:text text)))))
-
-
-(defn js-alert-error
-  "Render error message as a JavaScript alert"
-  [text] {:post [(types/error-text? %)]
-          :pre  [(types/error-text? text)]}
-  (assoc
-    text :text
-    (format "<script language='javascript'>alert(\"ERROR: %s\");</script>"
-            (reduce str (map #(if (#{\' \"} %) (str \\ %) %) (:text text))))))
+;(defn text-error
+;  "Render error message as regular text"
+;  [text] {:post [(types/error-text? %)]
+;          :pre  [(types/error-text? text)]}
+;  (assoc text :text (str "ERROR: " (:text text))))
+;
+;
+;(defn html-error
+;  "Render error message as HTML snippet"
+;  [text] {:post [(types/error-text? %)]
+;          :pre  [(types/error-text? text)]}
+;  (assoc
+;    text :text
+;    (format "<pre style='color: white; background-color: red'>ERROR: %s</pre>"
+;            (util/html-entities (:text text)))))
+;
+;
+;(defn js-alert-error
+;  "Render error message as a JavaScript alert"
+;  [text] {:post [(types/error-text? %)]
+;          :pre  [(types/error-text? text)]}
+;  (assoc
+;    text :text
+;    (format "<script language='javascript'>alert(\"ERROR: %s\");</script>"
+;            (reduce str (map #(if (#{\' \"} %) (str \\ %) %) (:text text))))))
 
 
 (def ^{:dynamic true
@@ -40,7 +40,8 @@
   of the message. You may like to override this to raise a
   platform-specific exception or to render for a specific
   environment (eg. web.)"}
-  *error* text-error)
+  *error* (fn [& args]
+            (assert (not "*error* rebound to error-handler")))) ;text-error
 
 
 (defn common-error
@@ -57,5 +58,3 @@
 (defn parse-error   [text] (common-error "Parse" text))
 (defn compile-error [text] (common-error "Compile" text))
 (defn render-error  [text] (common-error "Render" text))
-
-